moved Storage and added StorageInterface
